requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
channels==4.1.0
drf-yasg==1.21.8
whitenoise==6.8.2
Pillow==10.4.0
